Fix `pixel_values` error in torchtune service.
`pixel_values` and `image_grid_thw` are added to the micro batches as they are created (as [None] * num_sequences) if there are none, but torchtune expects only objects containing tensors when it calls `batch_to_device`. Maybe multimodal isn't supported on torchtune, but in that case, multimodal batch items shouldn't be added to the batch.